payroll negative; some progress on db-side validation

+++ src/main/java/edu/njit/cs631/fitness/service/api/ClazzAdministrationService.java
@@ -7,7 +7,7 @@ import edu.njit.cs631.fitness.data.entity.Clazz;
 public interface ClazzAdministrationService {
     void generateRandomClasses();
     void generateRandomRegistrations();
-	Clazz createClass(Integer exerciseId, Integer instructorId, Integer roomId, LocalDateTime start, Double duration);
+	Clazz createClass(Integer exerciseId, Integer instructorId, Integer roomId, LocalDateTime start, Integer duration);
     void deleteClazz(Integer clazzId);
 	void registerUserForClass(Integer userId, Integer clazzId);
     void deregisterUserForClass(Integer userId, Integer clazzId);
